Minor update, bug in basic part fixed (np.float32)

+++ source/_static/code/en/basic/example/numpy_manual_grad.py
@@ -1,7 +1,7 @@
 import numpy as np
 
-X_raw = np.array([2013, 2014, 2015, 2016, 2017])
-y_raw = np.array([12000, 14000, 15000, 16500, 17500])
+X_raw = np.array([2013, 2014, 2015, 2016, 2017], dtype=np.float32)
+y_raw = np.array([12000, 14000, 15000, 16500, 17500], dtype=np.float32)
 
 X = (X_raw - X_raw.min()) / (X_raw.max() - X_raw.min())
 y = (y_raw - y_raw.min()) / (y_raw.max() - y_raw.min())
@@ -11,11 +11,11 @@ a, b = 0, 0
 num_epoch = 10000
 learning_rate = 1e-3
 for e in range(num_epoch):
-    # 手动计算损失函数关于自变量（模型参数）的梯度
+    # Calculate the gradient of the loss function with respect to arguments (model parameters) manually.
     y_pred = a * X + b
     grad_a, grad_b = (y_pred - y).dot(X), (y_pred - y).sum()
 
-    # 更新参数
+    # Update parameters.
     a, b = a - learning_rate * grad_a, b - learning_rate * grad_b
 
 print(a, b)
